Rendering type filter widgets dynamically
